ansible: comment out vim plugins installation
My vim config will fail to be loaded in the first init to install the
plugins, so it needs manual confirmation.

There is 2 ways to solve this, either manually clonning each plugin and
using neovim builtin plugin manager or using the -Es mode in neovim,
this mode will supress errors.